Stand_app_1/requirements.txt
Flask==1.1.1
mysql-connector-python==8.0.18
PyYAML==3.12
requests==2.18.4
urllib3==1.22

